+++ bindgen/src/gen_kotlin_multiplatform/compounds.rs
@@ -7,7 +7,7 @@
 use uniffi_bindgen::backend::{Literal, Type};
 use uniffi_bindgen::ComponentInterface;
 
-use super::{AsCodeType, CodeType};
+use super::{AsCodeType, CodeType, Config};
 
 #[derive(Debug)]
 pub struct OptionalCodeType {
@@ -38,10 +38,12 @@ impl CodeType for OptionalCodeType {
         )
     }
 
-    fn literal(&self, literal: &Literal, ci: &ComponentInterface) -> String {
+    fn literal(&self, literal: &Literal, ci: &ComponentInterface, config: &Config) -> String {
         match literal {
             Literal::None => "null".into(),
-            Literal::Some { inner } => super::KotlinCodeOracle.find(&self.inner).literal(inner, ci),
+            Literal::Some { inner } => super::KotlinCodeOracle
+                .find(&self.inner)
+                .literal(inner, ci, config),
             _ => panic!("Invalid literal for Optional type: {literal:?}"),
         }
     }
@@ -76,7 +78,7 @@ impl CodeType for SequenceCodeType {
         )
     }
 
-    fn literal(&self, literal: &Literal, _ci: &ComponentInterface) -> String {
+    fn literal(&self, literal: &Literal, _ci: &ComponentInterface, _config: &Config) -> String {
         match literal {
             Literal::EmptySequence => "listOf()".into(),
             _ => panic!("Invalid literal for List type: {literal:?}"),
@@ -121,7 +123,7 @@ impl CodeType for MapCodeType {
         )
     }
 
-    fn literal(&self, literal: &Literal, _ci: &ComponentInterface) -> String {
+    fn literal(&self, literal: &Literal, _ci: &ComponentInterface, _config: &Config) -> String {
         match literal {
             Literal::EmptyMap => "mapOf()".into(),
             _ => panic!("Invalid literal for Map type: {literal:?}"),