+++ bindgen/src/templates/macros.kt
@@ -165,7 +165,7 @@
         {{ arg.name()|var_name }}: {{ arg|type_name(ci) }}
 {%-     if is_decl %}
 {%-         match arg.default_value() %}
-{%-             when Some with(literal) %} = {{ literal|render_literal(arg, ci) }}
+{%-             when Some with(literal) %} = {{ literal|render_literal(arg, ci, config) }}
 {%-             else %}
 {%-         endmatch %}
 {%-     endif %}